revisit dpi helpers
